add unit tests for wildlife
In order to add tests, this commit refactors the resources by removing
all business logic from the actual resource. It is thus moved to a
manager file. This handles all communication with the datastore. An
additional refactor is to improve configuration. This makes the db
instance injectable into the managers so that they can be mocked
at a later point in time. Due to this improved configuration, the
command to start the app has changed -- check out the readme for
details